moved splash components into a layout component
