+++ frontend/src/components/index.ts
@@ -1,6 +1,5 @@
 export * from './layout';
 export * from './table';
 export * from './ui';
-export * from './header';
 export * from './search';
 export * from './filters';
