+++ frontend/src/components/layout/constants.ts
@@ -1,0 +1,15 @@
+export const landAcknowledgement =
+  'The B.C. Public Service acknowledges the territories of First Nations around B.C. and is grateful to carry out our work on these lands. We acknowledge the rights, interests, priorities, and concerns of all Indigenous Peoples - First Nations, Métis, and Inuit - respecting and acknowledging their distinct cultures, histories, rights, laws, and governments.';
+
+export const footer = {
+  contactLink: {
+    name: 'Call, email, or text us, ',
+    url: 'https://www2.gov.bc.ca/gov/content/home/get-help-with-government-services',
+  },
+  serviceCentreLink: {
+    name: ' find a service centre',
+    url: 'https://www2.gov.bc.ca/gov/content/home/services-a-z',
+  },
+  tdMessage:
+    'We can help in over 120 languages and through Telephone Device For The Deaf (TDD). ',
+};
