+++ frontend/src/components/layout/private/Footer.tsx
