+++ frontend/src/components/layout/private/Layout.tsx
@@ -1,8 +1,8 @@
 import type { ReactElement } from 'react';
-import { Header } from '../header';
-import { APP_NAME } from '../../common';
-import { Footer } from '.';
+import { APP_NAME } from '../../../common';
 import { useKeycloak } from '@react-keycloak/web';
+import { Footer } from './Footer';
+import { Header } from './Header';
 
 export const Layout = ({ children }: { children: ReactElement }) => {
   const { keycloak } = useKeycloak();
@@ -10,15 +10,8 @@ export const Layout = ({ children }: { children: ReactElement }) => {
   return (
     <div className="min-h-screen flex flex-col">
       <Header appName={APP_NAME} authenticated={keycloak.authenticated} />
-      <div
-        className={[
-          'w-full  mt-12 ',
-          keycloak.authenticated ? ' bg-white' : ' bg-primaryBlue',
-        ].join(', ')}
-      >
-        {children}
-      </div>
-      {keycloak.authenticated && <Footer />}
+      <div className="w-full  mt-12 bg-white">{children}</div>
+      <Footer />
     </div>
   );
 };
