+++ frontend/src/components/layout/private/index.ts
@@ -1,0 +1,1 @@
+export * from './Layout';
