+++ frontend/src/components/layout/public/index.ts
@@ -1,0 +1,1 @@
+export * from './Layout';
