+++ frontend/src/pages/index.ts
@@ -1,4 +1,4 @@
 export { default as Dashboard } from './dashboard/Dashboard';
 export { default as Profile } from './profile/Profile';
 export { default as NotFound } from './NotFound';
-export { default as SplashPage } from './splash/SplashPage';
+export { default as SplashPage } from './SplashPage';
