--- frontend/src/pages/splash/constants.ts
@@ -1,32 +1,0 @@
-export const main = {
-  title: 'Welcome to CORE Team',
-  banner: '',
-  bannerLink: {
-    name: 'here ',
-    url: 'https://submit.digital.gov.bc.ca/app/form/submit?f=a6e29c0d-8d3b-4b8d-adfb-e025f150e3cd',
-  },
-  subtitle: 'Coordinated Operation Response in Emergencies',
-  description: {
-    pt1: 'This is the staffing system (previously known as TEAMS – Temporary Emergency Assignment Management System) that maintains a roster of employees from numerous ministries in the BC Government who have a combination of interest, skills, and experience to work within EMCR operation centres and BCWS fire centres to support communities in emergency management and wildfire response across the province.',
-  },
-  login: {
-    title: 'Login',
-    description: 'Use your IDIR to access the CORE member database.',
-  },
-};
-
-export const landAcknowledgement =
-  'The B.C. Public Service acknowledges the territories of First Nations around B.C. and is grateful to carry out our work on these lands. We acknowledge the rights, interests, priorities, and concerns of all Indigenous Peoples - First Nations, Métis, and Inuit - respecting and acknowledging their distinct cultures, histories, rights, laws, and governments.';
-
-export const footer = {
-  contactLink: {
-    name: 'Call, email, or text us, ',
-    url: 'https://www2.gov.bc.ca/gov/content/home/get-help-with-government-services',
-  },
-  serviceCentreLink: {
-    name: ' find a service centre',
-    url: 'https://www2.gov.bc.ca/gov/content/home/services-a-z',
-  },
-  tdMessage:
-    'We can help in over 120 languages and through Telephone Device For The Deaf (TDD). ',
-};
