--- frontend/src/pages/splash/index.ts
@@ -1,2 +1,0 @@
-export * from './SplashPage';
-export * from './constants';
